fix: search profiles & publications
