--- pages/search.tsx
@@ -1,68 +1,0 @@
-import { useCallback, useRef, useState } from 'react'
-import Link from 'next/link'
-//import styles from './search.module.css'
-
-export default function Search() {
-
-  const searchRef = useRef(null)
-  const [query, setQuery] = useState('')
-  const [active, setActive] = useState(false)
-  const [results, setResults] = useState([])
-
-  const searchEndpoint = (query:any) => `/api/search?q=${query}`
-
-  const onChange = useCallback((event:any) => {
-    const query = event.target.value;
-    setQuery(query)
-    if (query.length) {
-      fetch(searchEndpoint(query))
-        .then(res => res.json())
-        .then(res => {
-          setResults(res.results)
-        })
-    } else {
-      setResults([])
-    }
-  }, [])
-
-  const onFocus = useCallback(() => {
-    setActive(true)
-    window.addEventListener('click', onClick)
-  }, [])
-
-  const onClick = useCallback((event:any) => {
-    // if (searchRef.current && !searchRef.current.contains(event.target)) {
-    //   setActive(false)
-    //   window.removeEventListener('click', onClick)
-    // }
-  }, [])
-
-  return (
-    <div
-      //className={styles.container}
-      ref={searchRef}
-    >
-      <input
-        //className={styles.search}
-        onChange={onChange}
-        onFocus={onFocus}
-        placeholder='Search something'
-        type='text'
-        value={query}
-      />
-      { active && results.length > 0 && (
-        <ul //className={styles.results}
-        >
-          {results.map(({ id, title }) => (
-            <li //className={styles.result}
-             key={id}>
-              <Link href="/posts/[id]" as={`/posts/${id}`}>
-                <a>{title}</a>
-              </Link>
-            </li>
-          ))}
-        </ul>
-      ) }
-    </div>
-  )
-}
